juk/mpris2: Use SHA1 for track ID instead of base32.
This does mean that if we ever have to figure out a track from its ID
that JuK will have to do some work to figure out which track
corresponded to that track ID, but we don't need to worry about that
yet.

In the meantime this ensures that the DBus object path does not become
too long for the spec.