Add more-accurate accessors for track length info.
Our MPRIS implementation had an inaccurate length calculation because it
was assuming it was working with units of ms from the PlayerManager but
was in reality receiving seconds.

Phonon does export ms though, and many other methods of PlayerManager
work with ms. For the sake of consistency we add millisecond-based
methods (add instead of replacing due to the existing methods being part
of the public DBus API), which fixes the calculation as well.