mpris/juk Fix the CanPause property
From the spec:

"Note that this is an intrinsic property of the current track: its value
should not depend on whether the track is currently paused or playing.
In fact, if playback is currently paused (and CanControl is true), this
should be true."

JuK is unlikely to ever play anything that is not pausable.