Fix crash when resizing folder list dialog.
Christoph Feck was able to pin this down to an endless recursion in Qt
layout code, apparently involving both scroll bars trying to determine
when to automatically appear (or not).

Text elision is already setup well for directory names so I'm opting to
disable the horizontal scroll bar (the dialog can be resized larger if
necessary).

Additionally I'm integrating a method to set a minimum width since
there's a word-wrapped text label in the dialog (which is known to
interfere rather badly with Qt layouts if a minimum width is not set).
In this case I don't employ a code change since there's no reason this
label needs to be word-wrapped. Forcing "plain text" mode just helps
ensure Qt doesn't try to treat it like wrappable rich text.

If this doesn't fix the crash please re-open, but works well for me
here.

BUG:306736
FIXED-IN:4.9.2