Support reading/writing Ogg Opus tags.
I say Ogg Opus since it will later be possible to store Opus audio
within a Matroska container, instead we support Opus in an Ogg container
(which is what is recommended/assumed by the reference tools and the
standard).

This does not add playback support if the underlying Phonon backend
doesn't support Opus.

I've tested by adjusting metadata fields of various Opus files I've
encoded.

The Taglib feature that supports this has not been released yet as far
as I can tell from git, so we simply check to see whether the header we
need was installed.