folder-scan: Support a list of folders to ignore when auto-scanning.
Feature initially authored by Tom Li (see the Review Request and bug for
more details).

The idea is that the Manage Folders dialog now will have a list of
folders that should not be automatically searched (which would override
the list of folders to scan on startup). This way you can avoid having
to completely change up your music layout on-disk just to keep JuK from
automatically grabbing songs you don't want or need managed from within
JuK.

I ended up having to rework a bit to fix some corner cases (such as the
directory watcher which might flag new files under an excluded directory
after JuK startup), and avoid accidentally removing the ability to
manually select files under these excluded directories for addition to
the collection.

If a user does want to remove files that are already in the collection
list, they can enable the "File Name (Full Path)" column and then use
the search bar to find the offending path name, and select and remove
the files that are shown.

Note for posterity that the "Fixed In" version specifies the next
KDE-wide Software Compilation release, not the internal JuK version.

Thanks again to Tom for drafting the initial patch and being
extraordinarily patient while I tried to find time to review, re-review,
and make the finishing touches.

GUI:
REVIEW:110273
FEATURE:319106
FIXED-IN:4.11