coverinfo: Do not crash if an MP3 file has no ID3 tag.
This rarely happens since taglib will simply create an ID3 tag if one
didn't exist before... but taglib can't do this if it can't find the
file, which now seems to occur for files with a non-UTF8 filename
encoding.

You see this as a 'Couldn't resolve the mime type of <<foo>> -- this
shouldn't happen' message on the console. In this situation JuK is able
to resolve the file to an appropriate TagLib::File subclass, but the
TagLib object is in an invalid state.

CoverInfo wasn't checking for this (it was assuming any TagLib::File*
that it was given was valid).

I took a look through the reported JuK crasher bugs and didn't see
anything relevant.

I'm not sure whether this will end up in the KDE SC 4.11 or 4.11.1.