Fix 'trying to add empty file warnings'.
Finally backtraced it... to a signal handler for PlayerManager.  Turns
out we get the Stopped state with an old state of Loading after all,
which was being handled in the else block which assumed
buffering/playback, and tried to read a (null) FileHandle as a result
since nothing was actually playing.