Use a less magical way of hooking widgets together in tag editor.
Before this, we used a convention involving object names assigned in the
.ui XML to figure out at runtime which QCheckBoxes were used to enable
which widgets for multi-item edit ("configuration by convention" before
Rails made it cool!).

But, I found it hard to understand years later so I'm trying for
something more direct.  Hopefully it will be easier to understand to
whoever reads this 10 years from now.