systemtray: Fix crash with scroll-to-next.
And crashes involving other action invocations.  We need to guard
against KSNI::action everywhere action() is called, not just in the
ctor.  A good Junior Job might be going through and renaming the
JuK-specific ActionCollection and action() functions to something a
little less likely to conflict with KDE or Qt namespace.