Fix build when tunepimp is available.
The port to KF5 didn't encompass tunepimp, which was outdated even in
KDE4 times.  This is just a simple build fix since we're in middle of
release tagging, but in the future we will also need to audit the dead
code that's not likely to be replaced by musicbrainz's current
recommendations for client libs.