Immediately update dynamic/search playlist items upon a change.
This is as opposed to recording the new list of PlaylistItems and then
effecting the change after draining the event loop (an attempt at
keeping GUI responsive if many updates are happening).

Otherwise it is possible to
- Generate a list of PlaylistItems to read from (when updating the
playlist) and push to end of event loop
- Delete some of those PlaylistItems due to already-queued events to
delete playlist items
- Finally try to use the list of PlaylistItems and run into a segfault
from trying to access a deleted PlaylistItem.

I think this should fix a crash when you delete multiple playlists.

BUG:389176
FIXED-IN:17.12.2