Bring back a visual indication of the playing item.
Turned out to be surprisingly more difficult than I thought, so
hopefully no additional bugs.

In particular some Phonon backends seem to insert a trip to a
StoppedState when changing the MediaObject's current source, so I had to
find a way around that to be able to call Playlist::stop() from
PlaylistCollection (which is what actually clears the visual indication
when playback stops).