Remove ancient Exec= key from .desktop file.
Should allow JuK to start on Wayland.  Should also allow arguments to be
passed to JuK (which in theory supports loading files from command
line).

I ask for URLs to be passed (%U) instead of files (%F) to avoid
over-eager file managers or desktops trying to download music files to a
temp directory and then deleting them once JuK is done.

GUI:
BUG:391354
FIXED-IN:18.04