mpris: Don't toggle playback with "Pause".
We already have a separate playPause action for this, both in the MPRIS2
DBus adaptor, and in PlayerManager.

This also fixes playback resuming when the computer is suspended (Plasma
sends the "Pause" command to each media player regardless of whether
it's playing, and JuK incorrectly treats that as a request to toggle
playback).

BUG:381271
FIXED-IN:18.04