Bring back inline tag editing.
This was a KDE 4 JuK feature that had been left disabled in the port to
Plasma 5/KF5. I finished the implementation here in response to a user
request.

Given the code changes it's technically best described as a new feature
(even if it used to work in KDE 4) so it probably won't show up until
the next release of KDE Applications, though it should be safe to
backport for interested distributions.

GUI:
BUG:395948
FIXED-IN:18.08