Set the tick interval, because the default value is 0
Summary:
With this change the time is correctly updated also when using
the VLC backend.
It worked for some reason with phonon-gstreamer, but it looks like
it's just by accident. In fact, according the documentation, the default
value of the property is 0:
https://api.kde.org/phonon/html/classPhonon_1_1MediaObject.html#ae0f0ab91c97a1f3495ed9467048bb81a

The choosen value (100) should allow for a fluid scrolling of the
progress element, without having a too small resolution, and
it seems to match the resolution of the gstreamer backend anyway.

BUG: 389907

Test Plan:
The time is updated on both the progress bar and the status bar with
both VLC and GStreamer Phonon backends.

Reviewers: mpyne

Reviewed By: mpyne

Differential Revision: https://phabricator.kde.org/D15222