PlayerManager: Always perform setup on construction.
The only reason we couldn't perform setup on construction was because of
some U/I setup code that couldn't happen when the player manager is
normally constructed.

But the U/I setup is simple enough to move to JuK, so do so. Now the
rest of PlayerManager is that much simpler.