Enable Qt support for High DPI icons.
As submitted by Guo Yunhe <i@guoyunhe.me> on kde-multimedia.

In my testing this only applies with QT_AUTO_SCREEN_SCALE_FACTOR
environment variable enabled, but does indeed allow for many (not all)
pixmaps to automatically load high-res versions (such as from
KIconLoader).

I consider this a bugfix and the support is non-invasive, so applying to
18.12 and will merge to master separately.

Note that the cover art and slider art is still suboptimal.