Fix the "Show Columns" menu to show correct state on startup.
The "Show Columns" menu would previously have a checkbox entered for
every possible column for the playlist shown by default at JuK startup,
even though most columns were not visible.

This turned out to be because the class that reads the column
configuration (Playlist::SharedSettings) was not being used to generate
the Show Columns menu options. Instead the Playlist's own (defaulted)
values were used.

At this point in execution we can't force SharedSettings->apply() to run
because the Playlist is not fully setup. Instead we expose the config
values and use those to generate the menu.