Force sorting on so that special playlist sort to top.
That is, so that special playlists hopefully sort to the top. Although
the sorting of special playlists (Collection List, Play Queue, etc.) has
worked for me since the port to KF5, I have a good report that it
doesn't for others.

The code itself appears correct but reviewing the Qt documentation
indicates that QTreeWidget (which is one of the base classes for the
list of playlists) doesn't sort by default.

If QTreeWidget doesn't sort, then the comparison function might not be
getting called at all, which would make it accident that it seems to
work for me.

CCBUG:402398