Port sort function for list of playlists to KF5.
Now that sorting is enabled, it turned out that the sort really was
broken for everybody (sorry!). The old Qt 3 and 4 "sort by subclassing
the list item" feature wasn't actually ported into the Qt 5
compatibilility list view. So although we were defining the compare
function, nothing was actually using it.

Things seemed to work fine on my local system but I suspect that's
because I had playlists from KDE 4 times that I copied over which would
have already sorted the special playlists to the beginning.

Since the old method doesn't work I've just added a hidden sort key
column for now. The "proper solution" is probably to use a
QSortFilterProxyModel but that will be some time away.

CHANGELOG:Restore proper sorting of the list of playlists.
BUG:402398
FIXED-IN:18.12.1