Continued search widget cleanup; remove needless KToolBar usage.
Continuing on from the work in commit fc6fecd6394c where Carl Schwan
noticed that the SearchLine class (used as a base search input element
within the PlaylistSplitter and within the Advanced Search playlist
dialog) didn't really need to be a QWidget.

The resulting improvement to the search line's appearance above the list
of music tracks came very close to looking like a no-overhead widget,
but there was still a 1-pixel padding on the left/right between the line
that wasn't present for the playlist tracks themselves.

The cause was the nearby SearchWidget class (used in the
PlaylistSplitter), which is a thin wrapper around SearchLine, and for
some reason is a KToolBar. This may have been intended at some point in
JuK's prehistory (e.g. to allow for dockable/movable search line) but at
this point the search line always shows up in the same spot and isn't
even parented into the toolwindow where a toolbar normally would go.

As a result this doesn't need to be anything other than a QWidget
either, and making this change allows the search bar to line up
pixel-perfect with the list of music tracks as it should.

GUI:
CHANGELOG:Removes unintended padding between search line and rest of interface.