test: Fix tagguessertest lookup error with ctest.
For whatever reason, running plain ctest or even "ninja test" on a
testing-enabled does not succeed in finding tagguessertest to run it,
probably because this test is located in the build directory's /bin
subfolder and the test metadata is incorrect.

However I was able to solve this just by using the existing ECM macro to
add a single test (https://api.kde.org/ecm/module/ECMAddTests.html)
directly, so that's what I did. ctest now runs and passes both tests
fine again.