slider: Simplify slider handling to fix seeking and clazy warning.
JuK playback/seeking has been sort of busted in some ways for all of
KF5. While trying to fix a few clazy warnings about sliderReleased being
overloaded to a parent Qt class, I realized that most of the code we
have for slider handling is needless, especially with Qt5 (I'm assuming
it made more sense in the Qt3 timeframe it descended from).

Rather than fighting Qt, just use the slider classes like they are meant
to be used. I keep the custom paint handler for now but that may go
later as well.

The hard part is keeping the PlayerManager (which is controlling the
position of the slider when the user is not) from fighting with the user
or overriding the user's inputs. I'm not sure it's perfect but it's much
better in my testing than it was before... although MPRIS is still
buggy.