clazy: Fix most remaining warnings.
I'm not going to worry about non-POD statics in an application, and I
can't fix the Q_PROPERTY warnings in Phonon. Other than that, all
default clazy warnings should be gone now.

Of note is the tagtransactionmanager.cpp setFile call, which is not just
a warning but an actual bug (good job clazy!).