Fix use of undefined behavior in status label.
The event filter doesn't check that the received event is a QMouseEvent
but it is possible (and actually happened) to receive other events from
Qt.

But in this case we don't even need an event filter, we're just handling
mouse events and Qt provides entirely usable methods for that already.