history: Enforce specific date format that sorts right.
The QTreeWidget we're using sorts by display text by default. This makes
the date column in the history playlist a bit useless to sort by in the
normal format.

The ISO date format is a bit too ugly though, so I manually implement
something more akin to RFC 3339. To properly fix this we should
implement a specific role in each item that can be sorted by the
underlying model. If the underlying model is a QStandardItemModel this
might already be easy but I don't have time now to investigate further.