ui: Focus the playlist search line edit on startup.
This helps avoid the new QToolButton in the status bar from being
highlighted for no reason and is more convenient for playing music when
you first start JuK up.

In the process, use the standard Qt focus proxy feature rather than
manually emulating the effect, though this needs me to enable C++14 to
use the nicer qOverload<> instead of QOverload.