playlistcollection: Minor cleanups, rename object().
So the PlaylistBox stuff exists in a sort of unusual class structure:

  * PlaylistCollection, a plain class just holding playlist info
  * PlaylistCollection::ActionHandler, used for non-GUI stuff that still
    needs signals/slots (i.e. action handling)
  * PlaylistBox, the actual GUI class with full-blown QWidgets and the
    rest.

This sort of makes sense (it's an early instance of the approach Qt
would take with Model/View in Qt4) but it was even confusing me again.
So I've tried to rename to make the roles more clear, removed "using
namespace" stuff that muddies the picture up more and did some minor
other cleanups while I was here.