systemtray: Cleanups and modernization. Also a timer bugfix.
The bug: When using the track popup announcement, a timer kicks off to
tell the popup to fade out. A second timer is used to incrementally fade
the popup a bit every few milliseconds until the popup can finally be
hidden. However the first timer was never stopped, and would kick off
the fadeout sequence again, over and over, only hidden to view since the
popup is no longer visible.

This might be related to the old bug 165899 (which I tried to fix a
different way). I believe this happened in the Qt3/4 transition since
the code seems to assume the first timer was a "single shot" timer, but
it could have been wrong forever.

CCBUG:165899