Improve track sequencing by removing the track sequencing classes.
This removes one of my first contributions to JuK :(

But it's worth it because the extra code is not worth the complexity,
seeing as how the job is really pretty simple in the first place, even
with album random play and randomized playback.

I believe this also fixes some bugs, including some longstanding ones.
Bug 417551 (being unable to drag and drop into Play Queue) had some
related work in a recent commit but was still broken until now.

BUG:100356
BUG:166711
BUG:302250
BUG:303901
BUG:336637
BUG:353259
BUG:404157
BUG:417551
FIXED-IN:21.04