playqueue: Use an appropriate icon for rebadged "remove playlist" action.
A JuK user pointed out that the Play Queue context menu shows "Hide"
instead of "Remove Playlist" on the context menu, but that the icon
still looks like a trash can, which is not appropriate for "Hide".

I wasn't sure of a great icon to use either, but "list-remove" at least
does not imply that the Play Queue is being tossed in the rubbish.

BUG:336250
FIXED-IN:21.04