Fix longstanding bug preventing JuK from reloading Play Queue on startup.
This feature is gated behind the hidden-by-default "Save Play Queue on
Exit" action so you'd need to add that to your toolbar for now if you
wish to test.

The hardest part was realizing we can't cast upwards to Playlist* before
we call the right operator>> to read the playlist data.