taglib: Intercept debug messages to mention offending filename.
A long time ago TagLib gained the ability to let the application
intercept debug messages. It seems intended to allow for directing the
messages somewhere other than stderr, but we can use it to take care of
something that has long annoyed me:

TagLib will give messages indicating that there's an error in the file
it's reading, but without being clear on what file that is. While this
can be OK if the message is at the same time as an event like playing
the next track, it's more difficult in situations like scanning a
folder.