Bugfix to prevent JuK hang when new music items detected on startup.
Ran into something just now where JuK will hang forever waiting for a
lock at startup.

It turns out this happens if you've already had cached music items
before *and* have new music items in a watched folder.

In this case, the process of loading cached items will announce that the
cache has completed being loaded while the CollectionList items
dictionary is still locked.

Some of the code linked to this signal ends up scanning for new music
files and, if any are found, creates new CollectionList items for them.
However this also requires the item dictionary lock, and this leads to
deadlock.

I fix this by moving the re-scan process to be triggered from the event
loop (when the lock is no longer held). Probably been a bug for some
time. :(