Fix CLion issues
"if comment doesn't match the macro name
Declaration and assignment can be joined
Unpaired symbol: ')' seems to be missing
The word 'anyways' is informal American English. Did you mean anyway?
Use an instead of 'a' if the following word starts with a vowel sound, e.g. 'an article', 'an hour'.
Consider removing "of" to be more concise
Item is not correctly numbered. Expected item number 2, but got 1."