Clear temporary cover files when not in use
Instead of clearing these temporary files on next launch, remove them
immediately when they are no longer used. This saves a lot of space on
tmpfs, especially if the user has a large music collection. Also ensure
that temporary file names are unique to avoid conflicting names in a
multi-user environment.